full system

+++ ponderada4/produtos/services/produtos.py
@@ -5,6 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from repository.produtos import ProdutosRepository
 from models.produtos import Produto
 from schemas.produtos import Produto as ProdutosSchema
+from logs.logs import log_info, log_error, log_warning
 
 class ProdutosService:
     def __init__(self):
@@ -13,11 +14,16 @@ class ProdutosService:
     async def get(self, produto_id, db: AsyncSession):
         produto = await self.repository.get(produto_id, db)
         if produto is None:
+            log_error("Produto não encontrado")
             raise HTTPException(status_code=404, detail="Produto não encontrado")
         return produto
 
     async def get_all(self, db: AsyncSession):
-        return await self.repository.get_all(db)
+        produtos = await self.repository.get_all(db)
+        if not produtos:
+            log_warning("Nenhum produto encontrado")
+            raise HTTPException(status_code=404, detail="Nenhum produto encontrado")
+        return produtos
 
     async def add(self, produto : ProdutosSchema, db: AsyncSession):
         produto = await Produto(**produto.dict(db))
@@ -28,4 +34,9 @@ class ProdutosService:
         return self.repository.update(produto_id, produto, db)
 
     async def delete(self, produto_id, db: AsyncSession):
+        produto = await self.repository.get(produto_id, db)
+        if produto is None:
+            log_error("Produto não encontrado")
+            raise HTTPException(status_code=404, detail="Produto não encontrado")
+        
         return await self.repository.delete(produto_id, db)
