+++ ponderada4/usuarios/services/usuarios.py
@@ -5,6 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from repository.usuarios import UsuariosRepository
 from models.usuarios import Usuario
 from schemas.usuarios import Usuarios as UsuariosSchema
+from logs.logs import log_info, log_error, log_warning
 
 class UsuariosService:
     def __init__(self):
@@ -13,11 +14,16 @@ class UsuariosService:
     async def get(self, usuario_id, db: AsyncSession):
         usuario = await self.repository.get(usuario_id, db)
         if usuario is None:
+            log_error("Usuario não encontrado")
             raise HTTPException(status_code=404, detail="Usuario não encontrado")
         return usuario
 
     async def get_all(self, db: AsyncSession):
-        return await self.repository.get_all(db)
+        usuarios = await self.repository.get_all(db)
+        if not usuarios:
+            log_warning("Nenhum usuario encontrado")
+            raise HTTPException(status_code=404, detail="Nenhum usuario encontrado")
+        return usuarios
 
     async def add(self, usuario : UsuariosSchema, db: AsyncSession):
         usuario = await Usuario(**usuario.dict(db))
@@ -28,4 +34,8 @@ class UsuariosService:
         return self.repository.update(usuario_id, usuario, db)
 
     async def delete(self, usuario_id, db: AsyncSession):
+        usuario = await self.repository.get(usuario_id, db)
+        if usuario is None:
+            log_error("Usuario não encontrado")
+            raise HTTPException(status_code=404, detail="Usuario não encontrado")
         return await self.repository.delete(usuario_id, db)
