+++ ponderada4/wishlist/services/wishlist.py
@@ -5,6 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from repository.wishlist import WishlistRepository
 from models.wishlist import Wishlist
 from schemas.wishlist import Wishlist as WishlistSchema
+from logs.logs import log_info, log_error, log_warning
 
 class WishlistService:
     def __init__(self):
@@ -12,14 +13,15 @@ class WishlistService:
 
     async def get(self, wishlist_id, db: AsyncSession):
         wishlist = await self.repository.get(wishlist_id, db)
-        print(wishlist)
         if wishlist is None:
+            log_error("Wishlist não encontrada")
             raise HTTPException(status_code=404, detail="Wishlist não encontrada")
         return wishlist
 
     async def get_all(self, db: AsyncSession):
         wishlists = await self.repository.get_all(db)
         if not wishlists:
+            log_warning("Nenhuma wishlist encontrada")
             raise HTTPException(status_code=404, detail="Nenhuma wishlist encontrada")
     
 
@@ -32,4 +34,9 @@ class WishlistService:
         return self.repository.update(wishlist_id, wishlist, db)
 
     async def delete(self, wishlist_id, db: AsyncSession):
+        wishlist = await self.repository.get(wishlist_id, db)
+        if wishlist is None:
+            log_error("Wishlist não encontrada")
+            raise HTTPException(status_code=404, detail="Wishlist não encontrada")
+        
         return await self.repository.delete(wishlist_id, db)
